move pkg dedicated to the parser to parser pkg
