+++ builder/build.go
@@ -16,8 +16,8 @@ import (
 	"github.com/containerd/containerd/platforms"
 	"github.com/docker/distribution/reference"
 	"github.com/docker/dockerfile/dockerfile2llb"
-	"github.com/docker/dockerfile/dockerignore"
 	"github.com/docker/dockerfile/parser"
+	"github.com/docker/dockerfile/parser/dockerignore"
 	"github.com/docker/go-units"
 	controlapi "github.com/moby/buildkit/api/services/control"
 	"github.com/moby/buildkit/client/llb"
