+++ dockerfile2llb/convert_norunsecurity.go
@@ -4,7 +4,7 @@
 package dockerfile2llb
 
 import (
-	"github.com/docker/dockerfile/instructions"
+	"github.com/docker/dockerfile/parser/instructions"
 	"github.com/moby/buildkit/client/llb"
 )
