+++ dockerfile2llb/convert_ssh.go
@@ -1,7 +1,7 @@
 package dockerfile2llb
 
 import (
-	"github.com/docker/dockerfile/instructions"
+	"github.com/docker/dockerfile/parser/instructions"
 	"github.com/moby/buildkit/client/llb"
 	"github.com/pkg/errors"
 )
