+++ parser/instructions/bflag_test.go
