+++ parser/shell/envVarTest
