+++ parser/shell/equal_env_unix.go
