+++ parser/shell/equal_env_windows.go
